third_party/blink/web_tests/external/wpt/html/rendering/widgets/appearance/default-styles-expected.txt

This is a testharness.js-based test.
FAIL <input> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="text"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="TEXT"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="search"> assert_equals: -webkit-appearance expected "auto" but got "searchfield"
FAIL <input type="tel"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="url"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="email"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="password"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="date"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="month"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="week"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="time"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="datetime-local"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="number"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <input type="range"> assert_equals: -webkit-appearance expected "auto" but got "slider-horizontal"
FAIL <input type="color"> assert_equals: -webkit-appearance expected "auto" but got "square-button"
FAIL <input type="checkbox"> assert_equals: -webkit-appearance expected "auto" but got "checkbox"
FAIL <input type="radio"> assert_equals: -webkit-appearance expected "auto" but got "radio"
FAIL <input type="submit"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="reset"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="button"> assert_equals: -webkit-appearance expected "auto" but got "push-button"
FAIL <input type="unknowntype"> assert_equals: -webkit-appearance expected "auto" but got "textfield"
FAIL <select> assert_equals: -webkit-appearance expected "auto" but got "menulist"
FAIL <select multiple=""> assert_equals: -webkit-appearance expected "auto" but got "listbox"
FAIL <select size="2"> assert_equals: -webkit-appearance expected "auto" but got "listbox"
FAIL <button> assert_equals: -webkit-appearance expected "auto" but got "button"
FAIL <textarea> assert_equals: -webkit-appearance expected "auto" but got "textarea"
PASS <meter>
PASS <progress>
PASS <input type="hidden">
PASS <input type="HIDDEN">
PASS <input type="file">
PASS <input type="image">
PASS <div>
PASS <details>
PASS <summary>
PASS <video>
PASS <video controls="">
PASS <menuitem>
PASS <marquee>
PASS <keygen>
PASS <input> (namespace: null)
PASS <input> (namespace: http://www.w3.org/2000/svg)
PASS The html element
PASS The body element
Harness: the test ran to completion.
